services/worker/pyproject.toml
[tool.poetry]
name = "worker"
version = "0.1.0"
description = "Background worker for KindPath"
authors = ["KindPath Foundation"]

[tool.poetry.dependencies]
python = "^3.11"
celery = "^5.3.4"
kafka-python = "^2.0.2"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
numpy = "^1.24.3"
scipy = "^1.11.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"